+++ src/main/java/com/mindee/MindeeClient.java
@@ -9,6 +9,7 @@ import com.mindee.input.PageOptions;
 import com.mindee.pdf.PdfBoxApi;
 import com.mindee.pdf.PdfOperation;
 import com.mindee.pdf.SplitQuery;
+import com.mindee.v1.MindeeSettings;
 import com.mindee.v1.clientOptions.PredictOptions;
 import com.mindee.v1.clientOptions.WorkflowOptions;
 import com.mindee.v1.http.Endpoint;
