+++ src/main/java/com/mindee/v2/MindeeSettings.java
@@ -1,4 +1,4 @@
-package com.mindee;
+package com.mindee.v2;
 
 import java.util.Optional;
 import lombok.Builder;
@@ -9,13 +9,13 @@ import lombok.Getter;
  */
 @Getter
 @Builder
-public class MindeeSettingsV2 {
+public class MindeeSettings {
 
   private static final String DEFAULT_MINDEE_V2_API_URL = "https://api-v2.mindee.net/v2";
   private final String apiKey;
   private final String baseUrl;
 
-  public MindeeSettingsV2() {
+  public MindeeSettings() {
     this("", "");
   }
 
@@ -23,11 +23,11 @@ public class MindeeSettingsV2 {
     return Optional.ofNullable(apiKey);
   }
 
-  public MindeeSettingsV2(String apiKey) {
+  public MindeeSettings(String apiKey) {
     this(apiKey, "");
   }
 
-  public MindeeSettingsV2(String apiKey, String baseUrl) {
+  public MindeeSettings(String apiKey, String baseUrl) {
 
     if (apiKey == null || apiKey.trim().isEmpty()) {
       String apiKeyFromEnv = System.getenv("MINDEE_V2_API_KEY");
