+++ src/test/java/com/mindee/v2/MindeeSettingsTest.java
@@ -1,4 +1,4 @@
-package com.mindee;
+package com.mindee.v2;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
@@ -7,8 +7,8 @@ import org.junitpioneer.jupiter.SetEnvironmentVariable;
 public class MindeeSettingsTest {
 
   @Test
-  @SetEnvironmentVariable(key = "MINDEE_API_KEY", value = "abcd")
-  @SetEnvironmentVariable(key = "MINDEE_API_URL", value = "https://example.com")
+  @SetEnvironmentVariable(key = "MINDEE_V2_API_KEY", value = "abcd")
+  @SetEnvironmentVariable(key = "MINDEE_V2_API_URL", value = "https://example.com")
   void setEnvironmentVariablesAndEmptyParams() {
     MindeeSettings settings = new MindeeSettings("", "");
     Assertions.assertEquals("abcd", settings.getApiKey().orElse(""));
